Add the ablility to leave the clean copy behind for faster reprocessing.

+++ build_time/src/main.py
@@ -41,6 +41,8 @@ def main(args):
   parser.add_argument('output_dir', help='Output directory')
   parser.add_argument('--hinting', default=False, action='store_true',
                       help='Retain hinting if set, else strip hinting')
+  parser.add_argument('--reuse_clean', default=False, action='store_true',
+                      help='Reuse the "clean" file if possible')
   parser.add_argument('--log', default='WARNING',
                       help='Set the logging level; eg, --log=INFO')
 
@@ -53,22 +55,23 @@ def main(args):
   log = logging.getLogger()
   logging_handler = logging.StreamHandler(sys.stdout)
   logging_handler.setLevel(loglevel)
-  formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - '
-                                '%(message)s')
+  formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
   logging_handler.setFormatter(formatter)
   log.addHandler(logging_handler)
   log.setLevel(loglevel)
 
-  verbose = loglevel < logging.DEBUG
+  verbose = loglevel <= logging.DEBUG
   fontfile = cmd_args.fontfile
   # TODO(bstell) use Logger
   log.info('preprocess ' + cmd_args.fontfile)
   basename = os.path.basename(fontfile)
   filename, extension = os.path.splitext(basename)
   cur_time = datetime.datetime.now()
-  build_dir = ('tmp-%s-%04d-%02d-%02d-%02d-%02d-%02d.%d' %
-               (filename, cur_time.year, cur_time.month, cur_time.day,
-                cur_time.hour, cur_time.minute, cur_time.second, os.getpid()))
+  build_dir = 'tmp-%s' % filename
+  if not cmd_args.reuse_clean:
+    build_dir = ('%s-%04d-%02d-%02d-%02d-%02d-%02d.%d' %
+                 (build_dir, cur_time.year, cur_time.month, cur_time.day,
+                  cur_time.hour, cur_time.minute, cur_time.second, os.getpid()))
   output_dir = cmd_args.output_dir
   log.info('put results in ' + output_dir)
   try:
@@ -80,11 +83,15 @@ def main(args):
 
   cleanfile = filename + '_clean' + extension
   cleanfilepath = build_dir + '/' + cleanfile
-  log.debug('make cleaned up version: ' + cleanfilepath)
-  cleanup.cleanup(fontfile, cmd_args.hinting, cleanfilepath)
-  log.info(basename + '=' + str(os.path.getsize(fontfile)) + ', ' +
-           cleanfilepath + '=' + str(os.path.getsize(cleanfilepath)))
-  closure.dump_closure_map(cleanfilepath, build_dir)
+  cleanfile_exists = os.path.isfile(cleanfilepath)
+  if cmd_args.reuse_clean and cleanfile_exists:
+    log.debug('reuse cleaned up version: ' + cleanfilepath)
+  else:
+    log.debug('make cleaned up version: ' + cleanfilepath)
+    cleanup.cleanup(fontfile, cmd_args.hinting, cleanfilepath, verbose)
+    log.info(basename + '=' + str(os.path.getsize(fontfile)) + ', ' +
+             cleanfilepath + '=' + str(os.path.getsize(cleanfilepath)))
+    closure.dump_closure_map(cleanfilepath, build_dir)
   log.debug('start proprocess')
   preprocess = Preprocess(cleanfilepath, build_dir, verbose)
   log.debug('build base')
@@ -106,24 +113,27 @@ def main(args):
     log.error('jar command status: ' + str(status))
     return status
 
-  log.debug('move the files to the output directory')
-  mv_cmd = ('cd %s; mv %s %s %s' %
+  log.debug('cp the files to the output directory')
+  cp_cmd = ('cd %s; cp %s %s %s' %
             (build_dir, tachyfont_file, cleanfile, output_dir))
-  log.debug('mv_cmd: ' + mv_cmd)
-  status = os.system(mv_cmd)
-  log.debug('mv status ' + str(status))
+  log.debug('cp_cmd: ' + cp_cmd)
+  status = os.system(cp_cmd)
+  log.debug('cp status ' + str(status))
   if status:
-    log.error('mv status = ' + str(status))
+    log.error('cp status = ' + str(status))
     return status
 
-  log.debug('cleanup the build directory')
-  rm_cmd = ('rm -rf %s' % build_dir)
-  log.debug('rm_cmd: ' + rm_cmd)
-  status = os.system(rm_cmd)
-  log.debug('rm status ' + str(status))
-  if status:
-    log.error('rm status = ' + str(status))
-    return status
+  if cmd_args.reuse_clean:
+    log.debug('leaving the build directory: ' + build_dir)
+  else:
+    log.debug('cleanup the build directory')
+    rm_cmd = ('rm -rf %s' % build_dir)
+    log.debug('rm_cmd: ' + rm_cmd)
+    status = os.system(rm_cmd)
+    log.debug('rm status ' + str(status))
+    if status:
+      log.error('rm status = ' + str(status))
+      return status
 
   log.info('command status = ' + str(status))
   return status
